Phase 11 Step 11.1 follow-up: widen scale_factor from i32 to f64
REVIEW.md finding #178 was initially left as a deliberate, scope-
bounded decision (i32 preserved to avoid touching 40 demo call
sites). At the project owner's explicit follow-up request, both
PlatformConnection::scale_factor and WinitConnection::scale_factor
are widened to f64, returning winit's real per-window DPI value
directly instead of rounding it away. A workspace-wide grep before
the change confirmed the only real caller anywhere was smoke_test.rs's
own diagnostic print, so this touched one real call site, not forty.

+++ crates/tre-platform/src/lib.rs
@@ -110,8 +110,11 @@ impl PlatformConnection {
         }
     }
 
+    /// Real per-window DPI scale factor, at winit's own `f64` precision
+    /// (REVIEW.md finding #178: widened from a previously API-stability-
+    /// preserved `i32`, at the project owner's explicit direction).
     #[must_use]
-    pub fn scale_factor(&self, window: WindowId) -> i32 {
+    pub fn scale_factor(&self, window: WindowId) -> f64 {
         match self {
             Self::Wayland(c) => c.scale_factor(window),
             Self::X11(c) => c.scale_factor(window),
